Removed the space in place-helper & moved airplace package to player
